refactor(rendering): merge FeedAcceptHeader into FeedResponseFormat
Consolidates path extensions and HTTP Accept header parsing and negotiation rules into a single FeedResponseFormat module, removing the separate FeedAcceptHeader module.
